Updated frontend scaffolding to component hierarchy
* added app/js/components/
* * added boilerplate component in app/js/components/._example
* updated routing in app/js/on_config.js to use component hierarchy instead of view templates
* updated README.md documentation
