+++ app/index.html
@@ -7,12 +7,14 @@
       <title ng-bind="pageTitle"></title>
       <meta name="description" content="">
       <meta name="viewport" content="width=device-width">
-
+      
+      <link href="http://netdna.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css" rel="stylesheet">
       <link rel="stylesheet" href="css/main.css">
+
   </head>
   <body>
-
-      <div ui-view></div>
+       
+     <ui-view></ui-view>
 
       <script src="js/main.js"></script>
   </body>
