+++ app/js/components/app/app.component.js
@@ -1,0 +1,13 @@
+const AppComponent = {
+    templateUrl: 'app/app.tpl.html',
+    bindings: {},
+    controller: function () {
+    	
+    	
+    }
+}
+
+export default {
+	name: 'app',
+	obj: AppComponent
+};
