+++ app/js/components/app/app.tpl.html
@@ -1,0 +1,3 @@
+<nav-bar></nav-bar>
+<a ui-sref="app.home">Home</a>
+<ui-view></ui-view>
