+++ app/js/components/home/home.tpl.html
@@ -1,0 +1,12 @@
+
+
+<h1 class="heading -large">{{ $ctrl.title | ExampleFilter }}!</h1>
+
+<h3 class="heading -medium">Here is a fancy number served up courtesy of Angular: <span class="number-example">{{ $ctrl.number }}</span></h3>
+
+<img src="images/angular.png" height="100" />
+<img src="images/gulp.png" height="100" />
+<img src="images/browserify.png" height="100" />
+
+<hr/>
+<div example-directive title="WOW!" click-message="You clicked me!">Directive is not loaded.</div>
