+++ app/js/constants.js
@@ -3,4 +3,6 @@ const AppSettings = {
   apiUrl: '/api/v1'
 };
 
+
+
 export default AppSettings;
