--- app/views/directives/example.html
@@ -1,4 +1,0 @@
-<div class="example-directive">
-  <h1>Directive title: {{title}}</h1>
-  <p>This is an example of a directive, click me!</p>
-</div>
