+++ app_orig/results/results_test.js
@@ -1,0 +1,16 @@
+'use strict';
+
+describe('cognoma.view2 module', function() {
+
+  beforeEach(module('cognoma.view2'));
+
+  describe('view2 controller', function(){
+
+    it('should ....', inject(function($controller) {
+      //spec body
+      var view2Ctrl = $controller('View2Ctrl');
+      expect(view2Ctrl).toBeDefined();
+    }));
+
+  });
+});
