+++ app_orig/welcome/welcome.html
@@ -1,0 +1,59 @@
+
+
+<header class="container-fluid text-center">
+
+    <div class="page-header" >
+
+
+        <h1>Welcome to Cognoma!</h1>
+        <img src="images/Logo2Cognoma.jpg">
+        <p><i>Putting Machine Learning in the hands of Cancer Biologists.</i></p>
+    </div>
+
+</header>
+
+<div class="container-fluid">
+    <div class="row content">
+        <div class="well">
+        <center>
+            <button type="button" class="btn btn-info" data-toggle="collapse" data-target="#CognomaDescription">
+            What is Cognoma?</button>
+        </center>
+        <div id="CognomaDescription" class="collapse">
+            <br>
+            <center>
+            <p>Here we put some basic info about Cognoma, and maybe some links to more detailed information.</p>
+            <p>Blah, Blah, Blah!</p>
+            </center>
+        </div>
+        </div>
+    </div>
+
+    <div class="row content">
+        <!--<div class="well">-->
+            <form role="form">
+            <div class="col-sm-4"><center>
+                <button type="button" class="btn btn-default  btn-block">
+                    <span class="glyphicon glyphicon-log-in"></span> Log in
+                </button>
+                </center>
+            </div>
+            <div class="col-sm-4"><center>
+                <!--<button type="button" class="btn btn-default  btn-block">
+                    <span class="glyphicon glyphicon-log-in"></span> Anonymous Log in
+                </button>-->
+                <a class="btn btn-default btn-block" type="button" href="#!/main">
+                    <span class="glyphicon glyphicon-log-in"></span> Anonymous Log in</a>
+                </center>
+            </div>
+            <div class="col-sm-4"><center>
+                <button type="button" class="btn btn-default  btn-block">
+                    <span class="glyphicon glyphicon-user"></span> Sign Up
+                </button>
+                </center>
+            </div>
+            </form>
+        <!--</div>-->
+    </div>
+
+</div>
