+++ app_orig/welcome/welcome_test.js
@@ -1,0 +1,16 @@
+'use strict';
+
+describe('cognoma.welcome module', function() {
+
+  beforeEach(module('cognoma.welcome'));
+
+  describe('welcome controller', function(){
+
+    it('should ....', inject(function($controller) {
+      //spec body
+      var welcomeCtrl = $controller('welcomeCtrl');
+      expect(welcomeCtrl).toBeDefined();
+    }));
+
+  });
+});
