+++ package.json
@@ -36,7 +36,10 @@
   },
   "devDependencies": {
     "angular": "^1.5.0",
+    "angular-animate": "^1.5.8",
     "angular-mocks": "^1.3.15",
+    "angular-touch": "^1.5.8",
+    "angular-ui-bootstrap": "^2.1.3",
     "angular-ui-router": "^0.2.13",
     "babel-core": "^6.3.26",
     "babel-eslint": "^5.0.0",
